Add to_otu() method to SampleCollection * Moved to_otu from Classifications instance to SampleCollection * Moved magic_classifications_fetch from EnhancedSampleCollection to SampleCollection * Changed test_where_clauses test to pass with additional requests being generated
